+++ perl/FSM/Scheduler/ISF/Emitter/JSON.pm
@@ -29,6 +29,8 @@ sub report_hash($self, $ir) {
         clock          => $ir->{clock},
         reset          => $self->_reset_summary($ir->{reset}),
         watchdog       => $ir->{watchdog},
+        actor_phases   => $self->_actor_metadata_summary($ir, 'actor_phases'),
+        actor_stages   => $self->_actor_metadata_summary($ir, 'actor_stages'),
         actor_constants => $self->_actor_constant_summary($ir),
         port_count     => scalar(@{$ir->{ports}}),
         inputs         => scalar(grep { $_->{direction} eq 'input'  } @{$ir->{ports}}),
@@ -398,6 +400,17 @@ sub _is_runtime_wait_count_kind($count_kind) {
     return defined($count_kind) && $count_kind =~ /\Aruntime_/;
 }
 
+sub _actor_metadata_summary($self, $ir, $key) {
+    return [
+        map {
+            {
+                name => $_->{name},
+                body => _clone_report_value($_->{body} || []),
+            }
+        } @{$ir->{$key} || []}
+    ];
+}
+
 sub _transaction_loop_summary($self, $ir) {
     my @loops;
 
@@ -432,6 +445,12 @@ sub _transaction_loop_summary($self, $ir) {
     }
 
     return \@loops;
+}
+
+sub _clone_report_value($value) {
+    return [ map { _clone_report_value($_) } @$value ] if ref($value) eq 'ARRAY';
+    return { map { $_ => _clone_report_value($value->{$_}) } keys %$value } if ref($value) eq 'HASH';
+    return $value;
 }
 
 sub _transaction_stage_summary($self, $ir) {