+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -94,6 +94,8 @@ our @EXPORT_OK = qw(
     isf_public_interface_schedule_report_scheduled_fsm_shape
     isf_public_interface_schedule_report_source_shape
     isf_public_interface_schedule_report_state_count_shape
+    isf_public_interface_schedule_report_actor_phase_keys
+    isf_public_interface_schedule_report_actor_stage_keys
     isf_public_interface_schedule_report_actor_constant_keys
     isf_public_interface_schedule_report_storage_kind_values
     isf_public_interface_schedule_report_storage_optional_keys
@@ -210,6 +212,8 @@ sub build_isf_public_interface_contract {
         schedule_report_clock_domain_child_instance_keys => isf_public_interface_schedule_report_clock_domain_child_instance_keys(),
         schedule_report_clock_domain_crossing_endpoint_keys => isf_public_interface_schedule_report_clock_domain_crossing_endpoint_keys(),
         schedule_report_crossing_keys => isf_public_interface_schedule_report_crossing_keys(),
+        schedule_report_actor_phase_keys => isf_public_interface_schedule_report_actor_phase_keys(),
+        schedule_report_actor_stage_keys => isf_public_interface_schedule_report_actor_stage_keys(),
         schedule_report_actor_constant_keys => isf_public_interface_schedule_report_actor_constant_keys(),
         schedule_report_compile_issues_success_shape => isf_public_interface_schedule_report_compile_issues_success_shape(),
         schedule_report_compile_issue_keys => isf_public_interface_schedule_report_compile_issue_keys(),
@@ -413,6 +417,7 @@ sub build_isf_public_interface_contract {
             't/1247-isf-clock-domain-partition.t',
             't/1248-isf-rule-trigger-parameter-binding.t',
             't/1249-isf-activation-parameter-constants.t',
+            't/1252-isf-actor-phase-stage-report.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',
@@ -486,6 +491,8 @@ sub isf_public_interface_public_top_level_keys {
             schedule_report_scheduled_fsm_shape
             schedule_report_clock_shape
             schedule_report_watchdog_shape
+            schedule_report_actor_phase_keys
+            schedule_report_actor_stage_keys
             schedule_report_actor_constant_keys
             schedule_report_clock_domain_keys
             schedule_report_clock_domain_child_instance_keys
@@ -825,6 +832,8 @@ sub isf_public_interface_schedule_report_top_level_keys {
             clock
             reset
             watchdog
+            actor_phases
+            actor_stages
             actor_constants
             port_count
             inputs
@@ -1194,6 +1203,24 @@ sub isf_public_interface_schedule_report_actor_constant_keys {
     ];
 }
 
+sub isf_public_interface_schedule_report_actor_phase_keys {
+    return [
+        qw(
+            name
+            body
+        ),
+    ];
+}
+
+sub isf_public_interface_schedule_report_actor_stage_keys {
+    return [
+        qw(
+            name
+            body
+        ),
+    ];
+}
+
 sub isf_public_interface_schedule_report_clock_domain_keys {
     return [
         qw(
@@ -1512,6 +1539,8 @@ sub isf_public_interface_schedule_report_presence_key_family_map {
     return {
         schedule_report_top_level_keys => isf_public_interface_schedule_report_top_level_keys(),
         schedule_report_reset_keys => isf_public_interface_schedule_report_reset_keys(),
+        schedule_report_actor_phase_keys => isf_public_interface_schedule_report_actor_phase_keys(),
+        schedule_report_actor_stage_keys => isf_public_interface_schedule_report_actor_stage_keys(),
         schedule_report_actor_constant_keys => isf_public_interface_schedule_report_actor_constant_keys(),
         schedule_report_clock_domain_keys => isf_public_interface_schedule_report_clock_domain_keys(),
         schedule_report_clock_domain_child_instance_keys => isf_public_interface_schedule_report_clock_domain_child_instance_keys(),